Rename settings module to conf
As recommanded by django-appconf. This follows Django convention.
